Crate compiler_lib

Re-exports

pub use asciifile;

pub use diagnostics;

pub use lexer;

pub use firm_construction as firm;

pub use strtab;

pub use symtab;

pub use type_checking;

Modules

ast

backend

context
All state shared by the lexer, parser and semantic analysis phases.

optimization

print

ref_eq

timing
Measures time taken by single phases of the compiler.

visitor
The visitor module for the AST.

Structs

FirmContext
FirmContext is a singleton that represents the global state of libFIRM library configuration and FIRM-graph construction.

Enums

OutputSpecification