Struct compiler_lib::asciifile::position::Position

pub struct Position<'t> { /* fields omitted */ }

Methods

impl<'t> Position<'t>

pub fn at_file_start(file: &'t AsciiFile<'t>) -> Option<Self>

Create a new Position object pointing at the first character of a file. Returns None for empty files.

pub fn to_single_char_span(self) -> Span<'t>

pub fn byte_offset(&self) -> usize

pub fn file(&self) -> &AsciiFile<'t>

pub fn chr(&self) -> char

Get the character at this position

Guranteed to be within the ASCII range of UTF-8 that does not use the upper half of the bytes. Use byte() if you need a single byte representation instead.

pub fn byte(&self) -> u8

Get the byte at this position

For matching on the character or comparisons, you probably want chr() instead.

pub fn row(&self) -> usize

Return the row of the character's position within the file.

The row is zero based, meaning characters on the first line of the file are in row 0. To get the line number, add 1 to the return value.

pub fn line_number(&self) -> usize

Return the character's line number. Identical to row() + 1

pub fn column(&self) -> usize

Return the column of the characters position within the file.

The column is zero based, meaning the first characters of a line/row is positioned at column 0.

pub fn row_and_column(&self) -> (usize, usize)

Return the column of the characters position within the file.

The column is zero based, meaning the first characters of a line/row is positioned at column 0.

pub fn next(&self) -> Option<Position<'t>>

Get the position immediatly following this position or None if this is the last position in the file.

pub fn next_mut(self) -> Result<Self, Self>

The same as next() but in-place.

Fails if there is no next position and returns the unchanged position.

pub fn is_last(&self) -> bool

Check if we are at the end of the file (the last valid position/the last character of the file).

pub fn prev(&self) -> Option<Position<'t>>

Get the position immediatly following this position or None if this is the last position in the file.

pub fn prev_mut(self) -> Result<Self, Self>

The same as prev() but in-place.

Fails if there is no prev position and returns the unchanged position.

pub fn line(&self) -> Span<'t>

Get source code line containing the position.

The returned Span will include the trailing newline character of the line.

use compiler_lib::asciifile::{AsciiFile, Position, Span};

let file = AsciiFile::new(b"banana\napple\n\nkiwi").unwrap();

let lines = file
    .iter()
    .map(|position| {
        let line = position.line();
        (position.chr(), line.as_str().to_string())
    })
    .collect::<Vec<_>>();

let expected: Vec<(char, String)> = vec![
    ('b', "banana\n"),
    ('a', "banana\n"),
    ('n', "banana\n"),
    ('a', "banana\n"),
    ('n', "banana\n"),
    ('a', "banana\n"),
    ('\n', "banana\n"),
    ('a', "apple\n"),
    ('p', "apple\n"),
    ('p', "apple\n"),
    ('l', "apple\n"),
    ('e', "apple\n"),
    ('\n', "apple\n"),
    ('\n', "\n"),
    ('k', "kiwi"),
    ('i', "kiwi"),
    ('w', "kiwi"),
    ('i', "kiwi"),
]
.into_iter()
.map(|(c, s)| (c, s.to_string()))
.collect();

assert_eq!(expected, lines);

Windows style line endings are not considered. This means '\r' will be evaluated as a normal character without any special meaning.

Important traits for PositionIterator<'t>
impl<'t> Iterator for PositionIterator<'t> type Item = Position<'t>;

pub fn iter(&self) -> PositionIterator<'t>

Important traits for ReversePositionIterator<'t>
impl<'t> Iterator for ReversePositionIterator<'t> type Item = Position<'t>;

pub fn reverse_iter(&self) -> ReversePositionIterator<'t>