Struct compiler_lib::asciifile::spanned::Spanned

pub struct Spanned<'f, T> {
    pub span: Span<'f>,
    pub data: T,
}

Fields
span: Span<'f> data: T
Methods

impl<'f, T> Spanned<'f, T>

pub fn new(span: Span<'f>, value: T) -> Self

pub fn map<U, F>(self, f: F) -> Spanned<'f, U> where
F: FnOnce(T) -> U,

pub fn map_ref<U, F>(&self, f: F) -> Spanned<'f, U> where
F: FnOnce(&T) -> U,

pub fn combine_with<Rhs, Res, F>(
self,
rhs: Spanned<'f, Rhs>,
f: F
) -> Spanned<'f, Res> where
F: FnOnce(Self, Spanned<'f, Rhs>) -> Res,

Combine two Spanneds, aplying f to compute the contents of the new Spanned. There are also variants of this where LHS/Self, RHS, or both are boxed. These are necassary to avoid reallocation when calling this.

pub fn combine_with_boxed<Rhs, Res, F>(
self,
rhs: Box<Spanned<'f, Rhs>>,
f: F
) -> Spanned<'f, Res> where
F: FnOnce(Self, Box<Spanned<'f, Rhs>>) -> Res,

pub fn combine_boxed_with<Rhs, Res, F>(
self: Box<Self>,
rhs: Spanned<'f, Rhs>,
f: F
) -> Spanned<'f, Res> where
F: FnOnce(Box<Self>, Spanned<'f, Rhs>) -> Res,

pub fn combine_boxed_with_boxed<Rhs, Res, F>(
self: Box<Self>,
rhs: Box<Spanned<'f, Rhs>>,
f: F
) -> Spanned<'f, Res> where
F: FnOnce(Box<Self>, Box<Spanned<'f, Rhs>>) -> Res,