Module compiler_lib::debugging

This is a browser-based graphical debugger for FIRM graphs.

Usage

All interaction with the debugger is done through the breakpoint! macro. The macro can be called in two variants:

breakpoint!(label:&str,compiler_state:firm::Program|Graph) will print all graphs with each node labeled by its kind and id.

breakpoint!(label:&str,compiler_state:firm::Program|Graph, label_maker) can be used to highlight some nodes and to add additional information. label_maker can be a Fn(Node) -> Label or just a HashMap<Node,Label>.

So for example, to highlight the currently visited node and to add the current lattice value assigned to each node, you can write:

// assume a context similar to:
// let values :HashMap<Node,Tarval> = HashMap::new();
// while let Some(cur) = next_node()
breakpoint!("Constant Folding: iteration", self.graph, &|node| {
    let mut label = default_label(node);

    if let Some(tarval) = self.values.get(&node) {
        label = label.append(format!("\n{:?}", tarval));
    }

    if node == cur {
        label = label
            .style(Style::Filled)
            .fillcolor(X11Color::Blue)
            .fontcolor(X11Color::White);
    }

    label
});

Implementation

We lazily start a webserver serving the debugger the first time a breakpoint is encountered. We pause execution by blocking on a message send on a rendevouz-channel to the webserver. The message content is the current compiler state.

Structs

Breakpoint

CompiliationState

Debugger

DebuggerState

GUI

GuiThread

Enums

MsgToCompiler

MsgToGui

Constants

_IMPL_SERIALIZE_FOR_Breakpoint

_IMPL_SERIALIZE_FOR_CompiliationState

Statics

static_rocket_route_info_for_breakpoint
Rocket code generated static route info.

static_rocket_route_info_for_breakpoint_continue
Rocket code generated static route info.

static_rocket_route_info_for_breakpoint_list
Rocket code generated static route info.

static_rocket_route_info_for_snapshot_at_index
Rocket code generated static route info.

Functions

breakpoint

breakpoint_continue

breakpoint_list

check_updates

gui_thread

http_server

pause

rocket_route_fn_breakpoint
Rocket code generated wrapping route function.

rocket_route_fn_breakpoint_continue
Rocket code generated wrapping route function.

rocket_route_fn_breakpoint_list
Rocket code generated wrapping route function.

rocket_route_fn_snapshot_at_index
Rocket code generated wrapping route function.

snapshot_at_index

spawn_gui_thread