Module compiler_lib::linter

How to write an early lint:

use asciifile::Spanned;
use compiler_lib::linter::{AstLintPass, EarlyContext, LintContext, LintPass};
use diagnostics::{declare_lint, lint::LintArray, lint_array};
use parser::ast::*;

declare_lint!(
    pub MY_COOL_LINT,
    Warning,
    "do this the right way"
);

pub struct MyLintPass;

impl LintPass for MyLintPass {
    fn get_lints(&self) -> LintArray {
        lint_array!(MY_COOL_LINT)
    }
}

impl<'a, 'f> AstLintPass<'a, 'f> for MyLintPass {
    fn check_program(&mut self, cx: &EarlyContext<'_, '_>, program: &Spanned<'_, Program<'_>>) {
        cx.struct_lint(MY_COOL_LINT, program.span, "RIIR");
    }
}

How to write a late lint:

use asciifile::Spanned;
use compiler_lib::linter::{SemanticLintPass, LateContext, LintContext, LintPass};
use diagnostics::{declare_lint, lint::LintArray, lint_array};
use parser::ast::*;

declare_lint!(
    pub MY_COOL_LINT,
    Warning,
    "do this the right way"
);

pub struct MyLintPass;

impl LintPass for MyLintPass {
    fn get_lints(&self) -> LintArray {
        lint_array!(MY_COOL_LINT)
    }
}

impl<'a, 'f> SemanticLintPass<'a, 'f> for MyLintPass {
    fn check_program(&mut self, cx: &LateContext<'_, '_>, program: &Spanned<'_, Program<'_>>) {
        cx.struct_lint(MY_COOL_LINT, program.span, "RIIR");
    }
}

Don't forget to register your lint in one of

// Early lints
Linter::register_ast_passes();
// Late lints
Linter::register_semantic_passes();

Every lint module needs also to be registered in lints/mod.rs.

Structs

EarlyContext

LateContext

Linter

UnknownLint

Traits

AstLintPass

LintContext

LintPass

SemanticLintPass