Crate compiler_lib

Modules

asciifile

ast

context
All state shared by the lexer, parser and semantic analysis phases.

diagnostics
The diagnostics object controls the output of warnings and errors generated by the compiler during the lexing, parsing and semantic analysis phases. It also tracks the number of warnings and errors generated for flow control.

firm

lexer

parser

print

semantics

strtab
String table with zero-copy and amortised O(1) insert

symtab

type_checking
This module implements type checking on the AST. The type checking produces an instance of type_system::TypeSystem, which is later used by the firm module for IR graph construction.

visitor
The visitor module for the AST.